got rid of phoenix and using basic config with plug
