+++ README.md
@@ -59,8 +59,8 @@ It installs these packages (gocode gopkgs go-outline go-symbols guru gorename go
 
 To start your Phoenix app:
 
-  * Install elixir
+  * Install [elixir](http://elixir-lang.org/install.html)
   * Install dependencies with `mix deps.get`
-  * Start Phoenix endpoint with `mix phoenix.server`
+  * 
 
 Now you can visit [`localhost:4000`](http://localhost:4000) from your browser.
