+++ elixir/lib/parmyay.ex
@@ -1,0 +1,18 @@
+defmodule Parmyay do
+  @moduledoc """
+  Documentation for Parmyay.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> Parmyay.hello
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end
