+++ elixir/mix.exs
@@ -1,0 +1,32 @@
+defmodule Parmyay.Mixfile do
+  use Mix.Project
+
+  def project do
+    [
+      app: :parmyay,
+      version: "0.1.0",
+      elixir: "~> 1.4",
+      build_embedded: Mix.env == :prod,
+      start_permanent: Mix.env == :prod,
+      deps: deps()
+    ]
+  end
+
+  # Configuration for the OTP application
+  #
+  # Type "mix help compile.app" for more information
+  def application do
+    [
+      applications: [:cowboy, :plug],
+      extra_applications: [:logger]
+    ]
+  end
+
+  defp deps do
+    [
+      {:amnesia, "~> 0.2.7"},
+      {:cowboy, "~> 1.1"},
+      {:plug, "~> 1.3"}
+    ]
+  end
+end
