+++ elixir/test/parmyay_test.exs
@@ -1,0 +1,8 @@
+defmodule ParmyayTest do
+  use ExUnit.Case
+  doctest Parmyay
+
+  test "the truth" do
+    assert 1 + 1 == 2
+  end
+end
