--- elixirApi/test/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule Parmyay.PageControllerTest do
-  use Parmyay.ConnCase
-
-  test "GET /", %{conn: conn} do
-    conn = get conn, "/"
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
-  end
-end
