--- elixirApi/test/test_helper.exs
@@ -1,2 +1,0 @@
-ExUnit.start
-
