--- elixirApi/test/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule Parmyay.LayoutViewTest do
-  use Parmyay.ConnCase, async: true
-end
