--- elixirApi/test/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule Parmyay.PageViewTest do
-  use Parmyay.ConnCase, async: true
-end
