--- elixirApi/web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule Parmyay.PageController do
-  use Parmyay.Web, :controller
-
-  def index(conn, _params) do
-    render conn, "index.html"
-  end
-end
