--- elixirApi/web/router.ex
@@ -1,26 +1,0 @@
-defmodule Parmyay.Router do
-  use Parmyay.Web, :router
-
-  pipeline :browser do
-    plug :accepts, ["html"]
-    plug :fetch_session
-    plug :fetch_flash
-    plug :protect_from_forgery
-    plug :put_secure_browser_headers
-  end
-
-  pipeline :api do
-    plug :accepts, ["json"]
-  end
-
-  scope "/", Parmyay do
-    pipe_through :browser # Use the default browser stack
-
-    get "/", PageController, :index
-  end
-
-  # Other scopes may use custom stacks.
-  # scope "/api", Parmyay do
-  #   pipe_through :api
-  # end
-end
