--- elixirApi/web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule Parmyay.LayoutView do
-  use Parmyay.Web, :view
-end
