--- elixirApi/web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule Parmyay.PageView do
-  use Parmyay.Web, :view
-end
